🚚 Move `getPackageManagerData()` function to `src/utils/detect-package-manager.ts` file
